refactor(ui): extract audio + effects + effects-graph handler domains from main.js (S5 Stage 4)
Move the Phase 1 Audio handlers (samples / mixer / peaks / diagnostics /
output devices), the P3 effect chain with per-family parameter and meter
access (generic setEffectParameter, EQ, SmartBalance, Waveshaper, dynamics
visualization) and the graph-mode routing incl. graph-owned effect
instances (FXG.3-b/3-d/4-a) and the master graph verbatim out of
ui/main.js into ui/electron-main/{audio,effects,effects-graph}.js.
194 channels moved; names, behavior and defaults unchanged.

- audio.js — init({ safeHandler }); pulls userDataPath from ../runtimePaths.
- effects.js — init({ safeHandler, graphHandler }); defines and exports
  trackKey/masterKey. Chain/param mutations broadcast xleth:graph:changed,
  so graphHandler is injected alongside safeHandler.
- effects-graph.js — init({ safeHandler, graphHandler }); imports
  trackKey/masterKey from ./effects.

graphHandler / broadcastGraphChanged stay in main.js and are passed in as
deps. The dual chain-vs-graph effect families (nodeId chain API with
graph:changed broadcast vs effectInstanceId FXG API) were moved as-is; the
seam is now visible as effects.js vs effects-graph.js. Phase 0 handlers,
VST3, Export, Pattern/diagnostics, video frame output/shm and the WORLD
poll are untouched. main.js 2971 -> 2736 lines.

Verified: npm run build && npm run test:smoke -> 3/3 against both the
working tree and this exact committed main.js; vitest: 2478 passed, 2
failed = the pre-existing local-only mixer failures (uncommitted FxBadge
WIP), identical to the Stage 1-3 baseline.
